moved gadget script code to separate modules started implementing media player
