+++ src/helpers/obs_context_manager.py
@@ -61,3 +61,25 @@ def source_list_ar():
         yield source_list
     finally:
         obs.source_list_release(source_list)
+
+
+@contextmanager
+def media_source_ar(media_path: str, use_monitoring: bool):
+    media_source = obs.obs_source_create_private(
+        "ffmpeg_source", "Global Media Source", None
+        )
+    settings = data_ar()
+    obs.obs_data_set_string(settings, "local_file", media_path)
+    obs.obs_source_update(media_source, settings)
+    if use_monitoring:
+        obs.obs_source_set_monitoring_type(
+            media_source, obs.OBS_MONITORING_TYPE_MONITOR_AND_OUTPUT
+        )
+    else:
+        obs.obs_source_set_monitoring_type(
+            media_source, obs.OBS_MONITORING_TYPE_NONE
+        )
+    try:
+        yield media_source
+    finally:
+        obs.obs_source_release(media_source)
